pyproject.toml
[tool.poetry]
name = "sophia-ai-platform"
version = "1.0.0"
description = "Sophia AI Intelligence Platform"
authors = ["Sophia AI"]
readme = "README.md"
packages = [{include = "platform"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
aiohttp = "3.9.5"
asyncpg = "0.29.0"
fastapi = "0.111.0"
httpx = "0.26.0"
pydantic = "2.8.2"
python-dotenv = "1.0.1"
redis = "5.0.1"
uvicorn = {version = "0.30.1", extras = ["standard"]}
tenacity = "8.2.3"
pybreaker = "1.1.0"
cryptography = "42.0.8"
bcrypt = "4.1.2"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
passlib = {version = "1.7.4", extras = ["bcrypt"]}
python-multipart = "0.0.9"
structlog = "23.2.0"
sentry-sdk = {version = "1.45.0", extras = ["fastapi"]}
uvloop = "0.19.0"
httptools = "0.6.1"
safety = "^3.0.0"
bandit = "1.7.6"
PyJWT = {version = "2.8.0", extras = ["crypto"]}
aiofiles = "23.2.1"
sqlalchemy = "2.0.25"
qdrant-client = "1.7.0"
openai = "1.6.1"
ujson = "5.9.0"
numpy = "^1.26.0"
langgraph = "0.2.16"
langchain = "0.2.16"
langchain-core = "0.2.40"
langchain-community = "0.2.17"
psycopg2-binary = "2.9.9"
transformers = "4.36.2"
scikit-learn = "1.3.2"
tiktoken = "0.5.2"
astroid = "3.0.3"
prometheus-client = "0.19.0"
requests = "2.31.0"
agno = "*"
torch = "2.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-asyncio = "0.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"